Add discovery map
